PHP code to create unique buttons:
This code generates a unique ID for each button using the uniqid() function, which generates a unique identifier based on the current time in microseconds. The IDs are then used to create HTML buttons with unique IDs, which can be styled and manipulated using CSS and JavaScript.

You can modify this code to add more buttons or customize the HTML and CSS as needed. Keep in mind that the uniqid() function generates unique IDs, but they are not guaranteed to be completely random or secure. If you need more secure IDs, consider using a stronger random number generator or a dedicated secure ID generation library.